requirements.txt
python==3.10.4
beautifulsoup4==4.11.1
lxml==4.9.1
nltk==3.7
pandas==1.4.4
python-discord==1.7.3
requests==2.28.1
scikit-learn==1.1.2
